Update to accomodate helper context change.
Helper's context will be `undefined` soon, in preparation for a larger
refactoring of views + helpers that is hopefully going to land in Ember
1.12.

This PR makes sure that liquid-fire will function after that change.
`env.data.view` has been available since 1.10.0-beta.1, so this is safe
for all HTMLBars Ember versions.