new `movingSide` option on the fly-to transition
This lets you pick whether the old or new element will be the one that
animates. The motion is the same either way (always from the position &
size of the old element to the position and size of the new
element). But the content of the moving element will vary.