Adds a `media` contraint for matching transitions by media query
This is a simple but powerful feature that enables transitions to be
conditionally run based on a given media query. The query is set up as a
constraint and run with the `window.matchMedia` method.

Example:
```javascript
// Will run on small screen mobile devices
this.transition(
  this.toRoute('foo'),
  this.use('toLeft'),
  this.media('(max-width: 320px)')
);

// Will run on all other screen sizes
this.transition(
  this.toRoute('foo'),
  this.use('fade'),
  this.media('(min-width: 321px)')
);
```

Includes a polyfill for matchMedia that is installed through bower.