Don’t include DOM-specific JavaScript libraries
When running in FastBoot, Ember apps don’t need libraries that are used
for manipulating the DOM. Indeed, because those libraries often
implicitly assume they can access the DOM, they will crash when
evaluated or run in Node.js.

Previously, we were using a conditional to only evaluate the Velocity.js
library if `document` was not `undefined`.

That test was relatively brittle. Instead, this commit checks for the
existence of the `EMBER_CLI_FASTBOOT` environment variable and omits the
code entirely. This is not only more reliable, it reduces the size of
the FastBoot build.