Add configurable directions for liquid-spacer
Adds `growWidth` and `growHeight` boolean attributes to the Growable
mixin (and therefore to liquid-spacer). These default to `true` to
maintain existing behaviour.

Setting either to `false` will prevent liquid-spacer from animating
resizes in that direction. This is useful when you have a responsive
width element which you want to animate height transitions.

Fixes #392