Consider liquid-fire non-idle when a liquid-sync is still pending
During the time that we are blocked on a liquid-sync, liquid-fire doesn't have any running transitions. This can lead to confusing test failures, since you will try to wait for animations to finish when none have started yet, and your test will progress too soon.

This PR increments the running transitions count whenever we are blocked on a liquid-sync, which closes the gap and ensures you will really wait until an entire transition is completed.

I also made a related change that prevents confusing exceptions when a liquid-child is destroyed before it can notify its parent that it has rendereed.