locally scoped rules
You can now pass a `rules` option to liquid-outlet, liquid-bind, or liquid-if that contains a function just like the one normally exported from app/transitions.js. The local rules will completely replace the global rules. This is particularly useful for use in add components that want to package sophisticated animations