Use unrounded element dimensions
jQuery's outerWidth() and outerHeight() methods round an element's
dimensions. Since the measure function is used for setting width and
height on animated elements, it might happen that an element's size
is set to be smaller than it's actual size and information get missing.

getBoundingClientRect() returns unrounded values as used by Browsers
instead and solves that issue.