Changing travis & testem for better reliability
Phantom was causing lots of sporadic failures. Switching to Firefox (which is built into Travis) seems better.

I'm also switching back to a matrix of specific ember versions from versionCompatibility. I like that in theory but the way it compounds your probabiliyt of spurious failure and the total runtime made it not nice.