fix replacer
This only works when you run it inside the dummy app. In any real app it fails, because it will get the `addon` directory relative to the current working directory. It's especially confusing inside another addon, which has its own `addon` directory that will get incorporated into liquid-fire, replacing its own.

This is the source of errors like

> Could not find module `liquid-fire/velocity-ext` imported from `dummy/initializers/liquid-fire`